Changed implementation of key for xor operation
In prevision of creating a powershell empire module, the key to use for
the xor operation will be retreived in the registry. The powershell
empire module will store the key in
HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Control\Lsa in the "Key"
value. That way, a user will be able to specify the key they want to use
instead of hardcoding it in the DLL.